STORM-534:Add ServerInfo.java to storm nimbus server info, NimbusClient fetch nimbus info from zk
