+++ storm-core/src/clj/backtype/storm/cluster.clj
@@ -17,7 +17,8 @@
 (ns backtype.storm.cluster
   (:import [org.apache.zookeeper.data Stat ACL Id])
   (:import [org.apache.zookeeper KeeperException KeeperException$NoNodeException ZooDefs ZooDefs$Ids ZooDefs$Perms])
-  (:import [backtype.storm.utils Utils ServerInfo])
+  (:import [backtype.storm.utils Utils])
+  (:import [backtype.storm.generated ServerInfo])
   (:import [java.security MessageDigest])
   (:import [org.apache.zookeeper.server.auth DigestAuthenticationProvider])
   (:use [backtype.storm util log config])
@@ -319,7 +320,7 @@
         
       (nimbus-info
         [this]
-        (.instance ServerInfo (-> cluster-state
+        (Utils/instance  (-> cluster-state
             (get-data NIMBUS-SUBTREE false)
             maybe-deserialize)))
