+++ storm-core/src/jvm/backtype/storm/utils/Utils.java
@@ -54,6 +54,9 @@ import org.apache.zookeeper.ZooDefs;
 import org.apache.zookeeper.data.ACL;
 import org.apache.zookeeper.data.Id;
 import org.json.simple.JSONValue;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yaml.snakeyaml.Yaml;
@@ -62,6 +65,7 @@ import org.yaml.snakeyaml.constructor.SafeConstructor;
 import backtype.storm.Config;
 import backtype.storm.generated.ComponentCommon;
 import backtype.storm.generated.ComponentObject;
+import backtype.storm.generated.ServerInfo;
 import backtype.storm.generated.StormTopology;
 import backtype.storm.generated.AuthorizationException;
 
@@ -535,6 +539,21 @@ public class Utils {
         return delegate;
     }
 
+  public static ServerInfo instance(String jsonText) throws ParseException {
+    JSONParser parser = new JSONParser();
+    JSONObject jsonObject = (JSONObject) parser.parse(jsonText);
+    String host = (String) jsonObject.get("host");
+    Long port = (Long) jsonObject.get("port");
+    return new ServerInfo(host, port.intValue());
+  }
+
+  public static String toJsonString(ServerInfo serverInfo) {
+    JSONObject obj = new JSONObject();
+    obj.put("host", serverInfo.get_host());
+    obj.put("port", serverInfo.get_port());
+    return obj.toJSONString();
+  }
+
   public static ServerInfo getServerInfo(Map conf, String name) throws TTransportException {
     List<String> servers =
         (List<String>) conf.get(Config.STORM_ZOOKEEPER_SERVERS);
@@ -549,7 +568,7 @@ public class Utils {
       }
       byte[] zk_data = _curator.getData().forPath("/" + name);
       String jsonText = (String) Utils.deserialize(zk_data);
-      return ServerInfo.instance(jsonText);
+      return Utils.instance(jsonText);
     } catch (Exception e) {
       LOG.error(e.getMessage());
       throw new TTransportException(e);
