+++ storm-core/src/storm.thrift
@@ -155,10 +155,16 @@ struct SupervisorSummary {
   5: required string supervisor_id;
 }
 
+struct ServerInfo {
+  1: required string host;
+  2: required i32 port;
+}
+
 struct ClusterSummary {
   1: required list<SupervisorSummary> supervisors;
   2: required i32 nimbus_uptime_secs;
   3: required list<TopologySummary> topologies;
+  4: optional ServerInfo server_info;
 }
 
 struct ErrorInfo {
